+++ src/Node.cpp
@@ -12,7 +12,7 @@ Node::~Node()
 
 Node::Node(char name) : name(name) {}
 
-void Node::addPath(char ch, int dest)
+void Node::addPath(char ch, double dest)
 {
     myPath.insert(make_pair(ch, dest));
 }
